# specs/001-tharwa-platform-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tharwa API
  version: 1.0.0
  description: >
    Backend for Tharwa (ثروة). Public read endpoints are unauthenticated for MVP market browsing.
    Admin routes are under /admin/v1 (separate security scheme). Parallel EGP is not in MVP.
servers:
  - url: https://api.example.com
    description: Production (replace with real host)
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /v1/auth/register:
    post:
      tags: [Consumer]
      summary: Register consumer (password)
      operationId: consumerRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsumerRegisterRequest"
      responses:
        "201":
          description: Created + bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumerAuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/auth/social:
    post:
      tags: [Consumer]
      summary: Sign in or register with Google or Apple ID token
      operationId: consumerSocialLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsumerSocialLoginRequest"
      responses:
        "200":
          description: Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumerAuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/auth/login:
    post:
      tags: [Consumer]
      summary: Consumer password login
      operationId: consumerLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsumerLoginRequest"
      responses:
        "200":
          description: Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumerAuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/auth/forgot-password:
    post:
      tags: [Consumer]
      summary: Request password reset email
      operationId: consumerForgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsumerForgotPasswordRequest"
      responses:
        "200":
          description: Generic success (does not reveal whether email exists)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumerMessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/auth/reset-password:
    post:
      tags: [Consumer]
      summary: Set new password with reset token
      operationId: consumerResetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsumerResetPasswordRequest"
      responses:
        "200":
          description: New password set; returns auth tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumerAuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/auth/verify-email:
    post:
      tags: [Consumer]
      summary: Verify email with token from registration email
      operationId: consumerVerifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsumerVerifyEmailRequest"
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                type: object
                required: [verified]
                properties:
                  verified:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/auth/account:
    delete:
      tags: [Consumer]
      summary: Permanently delete signed-in consumer account
      description: >
        Password reconfirmation when the account has a password. OAuth-only accounts
        may omit password. Deletes watchlist, trade journal, portfolio positions,
        practice/sim account, price alerts, refresh tokens, and other personal data.
        Push devices linked to the account are unlinked (not deleted).
      operationId: consumerDeleteAccount
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsumerDeleteAccountRequest"
      responses:
        "204":
          description: Account deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    BadRequest:
      description: Validation or bad input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Unauthorized:
      description: Missing or invalid credentials / token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    TooManyRequests:
      description: Rate limited
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    ConsumerRegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
    ConsumerLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
          maxLength: 128
    ConsumerSocialLoginRequest:
      type: object
      required: [provider, idToken]
      properties:
        provider:
          type: string
          enum: [google, apple]
        idToken:
          type: string
          minLength: 20
          maxLength: 8192
    ConsumerForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ConsumerResetPasswordRequest:
      type: object
      required: [resetToken, newPassword]
      properties:
        resetToken:
          type: string
          minLength: 20
          maxLength: 512
        newPassword:
          type: string
          minLength: 8
          maxLength: 128
    ConsumerVerifyEmailRequest:
      type: object
      required: [verificationToken]
      properties:
        verificationToken:
          type: string
          minLength: 20
          maxLength: 512
    ConsumerDeleteAccountRequest:
      type: object
      properties:
        password:
          type: string
          minLength: 1
          maxLength: 128
    ConsumerMessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    ConsumerAuthResponse:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
        expiresIn:
          type: integer
        user:
          type: object
          required: [id, email]
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
...